ref bugs

+++ resources/views/components/action-buttons.blade.php
@@ -5,7 +5,7 @@
     @endrole
     @if(\Route::current()->getName() == 'users.index')
         @if(auth()->id() != $id)
-            {{ Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $id]]) }}
+            {{ Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $id], 'onsubmit' => 'return confirm("are you sure ?")']) }}
             {{ Form::button('<i class="fas fa-trash" aria-hidden="true"></i>',
                 ['class' => 'btn btn-warning btn-sm', 'type' => 'submit', 'style' => 'padding:0; background: none; border:none; color:#007bff;'] )  }}
             {{ Form::close() }}
